Added support to be in multiple channels.
This commit breaks compatibility with upstream, my changes was to
intrusive so I took the decision to breake compatibility.

This codebase allows the player to be in multiple channels and move
between them without parting from a channel.

I have removed the -o and -a from chanlist, they are integrated into
the normal listings. -p is replaced with the more verbose "hasplayer".